perf: iterate children using a for loop
Spread is considerably slower here than iterating through the converted
children and pushing them individually.

Additionally, spreading the args can result in a large number of params
which can cause other problems (sometimes `push` will throw).

I suspect the for loop is faster because `...convertChildren(...)` has
to iterate the full set in order to allocate it to an array in memory,
which it then has to iterate to spread as params. Whereas iterating
through one at a time means we can add the child to the block without
waiting for the entire iterable to finish.